+++ packages/kernel-test-local/src/sample-agent.e2e.test.ts
@@ -1,12 +1,12 @@
 import '@ocap/repo-tools/test-utils/mock-endoify';
 
+import { makeOllamaNodejsKernelService } from '@metamask/kernel-language-model-service/ollama/nodejs';
 import { makeConsoleTransport, Logger } from '@metamask/logger';
 import type { MakeAgentArgs, Agent } from '@ocap/kernel-agents';
 import { getMoonPhase } from '@ocap/kernel-agents/capabilities/examples';
 import { count, add, multiply } from '@ocap/kernel-agents/capabilities/math';
 import { makeJsonAgent } from '@ocap/kernel-agents/json';
 import { makeReplAgent } from '@ocap/kernel-agents-repl';
-import { makeOllamaNodejsKernelService } from '@ocap/kernel-language-model-service/ollama/nodejs';
 import { fetchMock } from '@ocap/repo-tools/test-utils/fetch-mock';
 import {
   afterAll,