+++ packages/kernel-test-local/src/vats/lms-chat-vat.ts
@@ -1,8 +1,8 @@
 import type { ERef } from '@endo/eventual-send';
+import { makeChatClient } from '@metamask/kernel-language-model-service';
+import type { ChatService } from '@metamask/kernel-language-model-service';
 import { makeDefaultExo } from '@metamask/kernel-utils/exo';
 import type { Logger } from '@metamask/logger';
-import { makeChatClient } from '@ocap/kernel-language-model-service';
-import type { ChatService } from '@ocap/kernel-language-model-service';
 
 /**
  * A vat that uses a kernel language model service to perform a chat completion