+++ packages/kernel-language-model-service/CHANGELOG.md
@@ -7,4 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Initial release as a published package (previously the private `@ocap/kernel-language-model-service`) ([#955](https://github.com/MetaMask/ocap-kernel/pull/955))
+
 [Unreleased]: https://github.com/MetaMask/ocap-kernel/